chore: remove deprecated `AnalyticalCard` component
BREAKING CHANGE: The `AnalyticalCard` component has been removed. Please use the `Card` component instead.
