--- packages/main/src/components/AnalyticalCard/AnalyticalCard.mdx
@@ -1,53 +1,0 @@
-import { ArgTypesWithNote, ControlsWithNote, DocsHeader, Footer } from '@sb/components';
-import SubcomponentsSection from '@sb/docs/SubcomponentsSection.md?raw';
-import { Canvas, Description, Markdown, Meta } from '@storybook/blocks';
-import '@ui5/webcomponents-icons/dist/person-placeholder.js';
-import '@ui5/webcomponents/dist/Link.js';
-import { AnalyticalCardHeader } from '../AnalyticalCardHeader';
-import { NumericSideIndicator } from '../NumericSideIndicator';
-import { MessageStrip } from '../../webComponents/MessageStrip/index';
-import { MessageStripDesign } from '../../enums/MessageStripDesign';
-import * as ComponentStories from './AnalyticalCard.stories';
-
-<Meta of={ComponentStories} subcomponents={{ AnalyticalCardHeader, NumericSideIndicator }} />
-
-<DocsHeader />
-
-<MessageStrip
-  design={MessageStripDesign.Warning}
-  children={
-    <>
-      This component is deprecated and will be removed with our next major release (v2.0.0)! Please use
-      <ui5-link href={'https://sap.github.io/ui5-webcomponents-react/?path=/docs/data-display-card'}>
-        Card
-      </ui5-link> instead.
-    </>
-  }
-/>
-
-## Example
-
-<Canvas of={ComponentStories.Default} />
-
-## Properties
-
-<ControlsWithNote of={ComponentStories.Default} />
-
-<br />
-<br />
-
-<Markdown>{SubcomponentsSection}</Markdown>
-
-## AnalyticalCardHeader
-
-<Description of={AnalyticalCardHeader} />
-
-<ArgTypesWithNote of={AnalyticalCardHeader} />
-
-## NumericSideIndicator
-
-<Description of={NumericSideIndicator} />
-
-<ArgTypesWithNote of={NumericSideIndicator} />
-
-<Footer />
